requirements.txt
tweepy>=4.14.0
openai>=1.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
requests>=2.31.0
pydantic>=2.0.0
tenacity>=8.2.0

